Introduce a streaming iterator, vastly improving read performance.
The streaming "iterator" returns references to its own buffer instead of
copying items to a new vector.

A `to_iter` method is provided to get a regular `Iterator` for
convenience, at the expense of performance.

Additionally, a `QueueItemType` is introduced with `Regular` and
`Truncated` types. This provides a mechanism that can be expanded with
`Removed` and `Large` in the future.

There may be some bugs here, so this won't be released until more
comprehensive tests are performed.